refactor: svg 파일을 assets 폴더로 이동
