+++ src/constants/index.js
@@ -1,5 +1,4 @@
 export * from './messages.js';
 export * from './httpMethods.js';
 export * from './ownerName.js';
-export * from './svg.js';
 export * from './url.js';
